# apps/api/openapi.internal.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/listings": {
      "post": {
        "operationId": "ListingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/listings/mine": {
      "get": {
        "operationId": "ListingsController_findMine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/listings/by-ref/{reference}": {
      "get": {
        "operationId": "ListingsController_findByReference",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lang",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/listings/{id}": {
      "get": {
        "operationId": "ListingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "ListingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/listings/{id}/view": {
      "post": {
        "operationId": "ListingsController_registerView",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/listings/{id}/call": {
      "post": {
        "operationId": "ListingsController_registerCall",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/listings/{id}/status": {
      "patch": {
        "operationId": "ListingsController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnerStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/agent-applications": {
      "get": {
        "operationId": "AdminAgentApplicationsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/admin/agent-applications/{id}/approve": {
      "post": {
        "operationId": "AdminAgentApplicationsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agent-applications/{id}/reject": {
      "post": {
        "operationId": "AdminAgentApplicationsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectAgentApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/listings/{id}/status": {
      "patch": {
        "operationId": "AdminListingsController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Avino API",
    "description": "API портала недвижимости Avino. Аутентификация: POST /auth/otp/request → /auth/otp/verify → { accessToken, refreshToken }; обновление — /auth/refresh; вход через Google — /auth/google; вход через Apple — /auth/apple. Bearer-токен передаётся в заголовке Authorization.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "TourWindowDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "pattern": "HHMM_RE"
          },
          "end": {
            "type": "string",
            "pattern": "HHMM_RE"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "CreateListingTranslationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string"
          },
          "address_note": {
            "type": "string"
          },
          "features_text": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "CreateListingDto": {
        "type": "object",
        "properties": {
          "transaction_type": {
            "type": "object"
          },
          "property_type": {
            "type": "object"
          },
          "original_language": {
            "type": "object"
          },
          "price": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "currency": {
            "type": "object"
          },
          "area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "lot_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "living_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "non_living_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "rooms": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "bathrooms": {
            "type": "number",
            "minimum": 0,
            "maximum": 99
          },
          "parking_type": {
            "type": "object"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 80
            }
          },
          "floor": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "is_basement": {
            "type": "boolean"
          },
          "total_floors": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "year_built": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "city_id": {
            "type": "string"
          },
          "district_id": {
            "type": "string"
          },
          "agency_id": {
            "type": "string"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "tours_enabled": {
            "type": "boolean"
          },
          "tour_windows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourWindowDto"
            }
          },
          "translation": {
            "$ref": "#/components/schemas/CreateListingTranslationDto"
          }
        },
        "required": [
          "transaction_type",
          "property_type",
          "original_language",
          "price",
          "currency",
          "translation"
        ]
      },
...
      "UpdateListingTranslationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255
          },
          "description": {
            "type": "string"
          },
          "address_note": {
            "type": "string"
          },
          "features_text": {
            "type": "string"
          }
        }
      },
      "UpdateListingDto": {
        "type": "object",
        "properties": {
          "transaction_type": {
            "type": "object"
          },
          "property_type": {
            "type": "object"
          },
          "price": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "currency": {
            "type": "object"
          },
          "area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "lot_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "living_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "non_living_area": {
            "type": "string",
            "pattern": "DECIMAL_2"
          },
          "rooms": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "bathrooms": {
            "type": "number",
            "minimum": 0,
            "maximum": 99
          },
          "parking_type": {
            "type": "object"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 80
            }
          },
          "floor": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "is_basement": {
            "type": "boolean"
          },
          "total_floors": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "year_built": {
            "type": "number",
            "minimum": 0,
            "maximum": 32767
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "city_id": {
            "type": "string"
          },
          "district_id": {
            "type": "string"
          },
          "agency_id": {
            "type": "string"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "tours_enabled": {
            "type": "boolean"
          },
          "tour_windows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourWindowDto"
            }
          },
          "translation": {
            "$ref": "#/components/schemas/UpdateListingTranslationDto"
          }
        }
      },
      "OwnerStatusDto": {
        "type": "object",
        "properties": {
          "action": {
            "enum": [
              "HIDE",
              "MARK_SOLD",
              "MARK_RENTED",
              "REACTIVATE"
            ],
            "type": "string"
          }
        },
        "required": [
          "action"
        ]
      },
...
      "RejectAgentApplicationDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
      "ModerateListingDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "APPROVE",
              "SEND_TO_DRAFT",
              "REJECT",
              "DELETE"
            ]
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "action"
        ]
      },
...
    }
  }
}