# apps/api/src/listings/dto/create-listing.dto.ts
import { Type } from 'class-transformer';
import {
  ArrayMaxSize,
  IsArray,
  IsBoolean,
  IsEnum,
  IsInt,
  IsLatitude,
  IsLongitude,
  IsNotEmpty,
  IsNumber,
  IsOptional,
  IsString,
  IsUUID,
  Matches,
  Max,
  MaxLength,
  Min,
  ValidateIf,
  ValidateNested,
} from 'class-validator';
import {
  Currency,
  Language,
  ParkingType,
  PropertyType,
  TransactionType,
} from '@prisma/client';
import { IsHalfStep } from '../../common/validation/is-half-step';

/**
 * Денежные/площадные поля — строки-Decimal, никогда float (ADR-002). До 12 цифр
 * целой части и до 2 дробных (`listings.price` — Decimal(14,2)).
 */
const DECIMAL_2 = /^\d{1,12}(\.\d{1,2})?$/;

/** Время в формате HH:MM (00:00 – 23:59). */
const HHMM_RE = /^([01]\d|2[0-3]):[0-5]\d$/;

/** PostgreSQL `SmallInt` — допустимый диапазон значений для rooms/floor/year_built. */
const SMALLINT_MAX = 32767;

/**
 * Авторский перевод объявления на `original_language` (API.md §7, ADR-005).
 * Сохраняется как `ListingTranslation` с `source=USER`, `is_auto_translated=false`;
 * машинные переводы (uz/ru/en) генерируются отдельной задачей после `ACTIVE`.
 */
export class CreateListingTranslationDto {
  @IsString()
  @IsNotEmpty()
  @MaxLength(255)
  title!: string;

  @IsOptional()
  @IsString()
  description?: string;

  @IsOptional()
  @IsString()
  address_note?: string;

  @IsOptional()
  @IsString()
  features_text?: string;
}

/** Окно доступного времени тура (общее). `start`/`end` — `HH:MM`. */
export class TourWindowDto {
  @Matches(HHMM_RE, { message: 'start must be HH:MM' })
  start!: string;

  @Matches(HHMM_RE, { message: 'end must be HH:MM' })
  end!: string;
}

/**
 * Тело запроса `POST /api/v1/listings` (TASK-050, API.md §7).
 *
 * Объявление создаётся на одном языке (`original_language`) со статусом `NEW`
 * и проходит moderation queue (CLAUDE.md §9). Медиа — отдельная задача; удобства
 * принимаются как `amenities` (ADR-0111). Имена свойств — snake_case ключи контракта
 * (как в остальных DTO).
 */
export class CreateListingDto {
  @IsEnum(TransactionType)
  transaction_type!: TransactionType;

  @IsEnum(PropertyType)
  property_type!: PropertyType;

  @IsEnum(Language)
  original_language!: Language;

  @Matches(DECIMAL_2, {
    message: 'price must be a decimal string with up to 2 fraction digits',
  })
  price!: string;

  @IsEnum(Currency)
  currency!: Currency;

  @IsOptional()
  @Matches(DECIMAL_2, {
    message: 'area must be a decimal string with up to 2 fraction digits',
  })
  area?: string;

  @IsOptional()
  @Matches(DECIMAL_2, { message: 'lot_area must be a decimal string with up to 2 fraction digits' })
  lot_area?: string;

  /** Жилая площадь, м² (мобилка #10; клиент показывает для дома/особняка). */
  @IsOptional()
  @Matches(DECIMAL_2, { message: 'living_area must be a decimal string with up to 2 fraction digits' })
  living_area?: string;

  /** Нежилая площадь, м² (кухня/санузлы/коридоры). */
  @IsOptional()
  @Matches(DECIMAL_2, { message: 'non_living_area must be a decimal string with up to 2 fraction digits' })
  non_living_area?: string;

  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(SMALLINT_MAX)
  rooms?: number;

  /** Санузлы, шаг 0.5 (1, 1.5, 2 …) — баглист мобилки #3. */
  @IsOptional()
  @IsNumber({ maxDecimalPlaces: 1 })
  @IsHalfStep()
  @Min(0)
  @Max(99)
  bathrooms?: number;

  @IsOptional()
  @IsEnum(ParkingType)
  parking_type?: ParkingType;

  @IsOptional()
  @IsArray()
  @ArrayMaxSize(50)
  @IsString({ each: true })
  @MaxLength(80, { each: true })
  amenities?: string[];

  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(SMALLINT_MAX)
  floor?: number;

  /** Цокольный этаж (баглист мобилки #4). При true клиент обычно шлёт floor: null. */
  @IsOptional()
  @IsBoolean()
  is_basement?: boolean;

  @IsOptional()
  @IsInt()
  @Min(0)
  @Max(SMALLINT_MAX)
  total_floors?: number;

  /**
   * Год постройки. ОБЯЗАТЕЛЕН для квартир/домов: категория «новостройка»
   * вычисляется из него (`?new_construction=true`). Может быть в будущем —
   * недострой («сдача в 2028», квартиру перепродают до сдачи дома).
   */
  @ValidateIf(
    (o: CreateListingDto) =>
      o.property_type === PropertyType.APARTMENT ||
      o.property_type === PropertyType.HOUSE ||
      o.year_built !== undefined,
  )
  @IsInt({ message: 'year_built is required for APARTMENT/HOUSE and must be an integer' })
  @Min(0)
  @Max(SMALLINT_MAX)
  year_built?: number;

  @IsOptional()
  @IsString()
  @MaxLength(500)
  address?: string;

  @IsOptional()
  @IsUUID()
  city_id?: string;

  @IsOptional()
  @IsUUID()
  district_id?: string;

  @IsOptional()
  @IsUUID()
  agency_id?: string;

  // Координаты — только из map-picker, никогда из EXIF фото (ADR-008). Источник
  // для PostGIS `location` (синхронизация — отдельная гео-задача).
  @IsOptional()
  @IsLatitude()
  latitude?: string;

  @IsOptional()
  @IsLongitude()
  longitude?: string;

  @IsOptional()
  @IsBoolean()
  tours_enabled?: boolean;

  @IsOptional()
  @IsArray()
  @ArrayMaxSize(6)
  @ValidateNested({ each: true })
  @Type(() => TourWindowDto)
  tour_windows?: TourWindowDto[];

  @ValidateNested()
  @Type(() => CreateListingTranslationDto)
  translation!: CreateListingTranslationDto;
}